Add an option to use the new I/O safety types and traits in std.
I/O safety is now [in Rust Nightly]; add a mode to io-lifetimes to use
std's types and traits. See the blurb in README.md for more details.

[in Rust Nightly]: https://github.com/rust-lang/rust/pull/87329
