+++ src/traits.rs
@@ -1,21 +1,20 @@
+#[cfg(not(io_lifetimes_use_std))]
 #[cfg(any(unix, target_os = "wasi"))]
-use crate::{BorrowedFd, OwnedFd};
+use crate::BorrowedFd;
+#[cfg(any(unix, target_os = "wasi"))]
+use crate::OwnedFd;
+#[cfg(not(io_lifetimes_use_std))]
 #[cfg(windows)]
-use crate::{BorrowedHandle, BorrowedSocket, HandleOrInvalid, OwnedHandle, OwnedSocket};
-#[cfg(unix)]
-use std::os::unix::io::{AsRawFd, FromRawFd, IntoRawFd};
-#[cfg(target_os = "wasi")]
-use std::os::wasi::io::{AsRawFd, FromRawFd, IntoRawFd};
+use crate::{BorrowedHandle, BorrowedSocket};
 #[cfg(windows)]
-use std::os::windows::io::{
-    AsRawHandle, AsRawSocket, FromRawHandle, FromRawSocket, IntoRawHandle, IntoRawSocket,
-};
+use crate::{OwnedHandle, OwnedSocket};
 
 /// A trait to borrow the file descriptor from an underlying object.
 ///
 /// This is only available on unix platforms and must be imported in order to
 /// call the method. Windows platforms have a corresponding `AsHandle` and
 /// `AsSocket` set of traits.
+#[cfg(not(io_lifetimes_use_std))]
 #[cfg(any(unix, target_os = "wasi"))]
 pub trait AsFd {
     /// Borrows the file descriptor.
@@ -23,6 +22,7 @@ pub trait AsFd {
     /// # Example
     ///
     /// ```rust,no_run
+    /// # #![cfg_attr(io_lifetimes_use_std, feature(io_safety))]
     /// use std::fs::File;
     /// # use std::io;
     /// use io_lifetimes::{AsFd, BorrowedFd};
@@ -35,6 +35,7 @@ pub trait AsFd {
 }
 
 /// A trait to borrow the handle from an underlying object.
+#[cfg(not(io_lifetimes_use_std))]
 #[cfg(windows)]
 pub trait AsHandle {
     /// Borrows the handle.
@@ -42,6 +43,7 @@ pub trait AsHandle {
     /// # Example
     ///
     /// ```rust,no_run
+    /// # #![cfg_attr(io_lifetimes_use_std, feature(io_safety))]
     /// use std::fs::File;
     /// # use std::io;
     /// use io_lifetimes::{AsHandle, BorrowedHandle};
@@ -54,6 +56,7 @@ pub trait AsHandle {
 }
 
 /// A trait to borrow the socket from an underlying object.
+#[cfg(not(io_lifetimes_use_std))]
 #[cfg(windows)]
 pub trait AsSocket {
     /// Borrows the socket.
@@ -69,6 +72,7 @@ pub trait IntoFd {
     /// # Example
     ///
     /// ```rust,no_run
+    /// # #![cfg_attr(io_lifetimes_use_std, feature(io_safety))]
     /// use std::fs::File;
     /// # use std::io;
     /// use io_lifetimes::{IntoFd, OwnedFd};
@@ -89,6 +93,7 @@ pub trait IntoHandle {
     /// # Example
     ///
     /// ```rust,no_run
+    /// # #![cfg_attr(io_lifetimes_use_std, feature(io_safety))]
     /// use std::fs::File;
     /// # use std::io;
     /// use io_lifetimes::{IntoHandle, OwnedHandle};
@@ -117,6 +122,7 @@ pub trait FromFd {
     /// # Example
     ///
     /// ```rust,no_run
+    /// # #![cfg_attr(io_lifetimes_use_std, feature(io_safety))]
     /// use std::fs::File;
     /// # use std::io;
     /// use io_lifetimes::{FromFd, IntoFd, OwnedFd};
@@ -134,6 +140,7 @@ pub trait FromFd {
     /// # Example
     ///
     /// ```rust,no_run
+    /// # #![cfg_attr(io_lifetimes_use_std, feature(io_safety))]
     /// use std::fs::File;
     /// # use std::io;
     /// use io_lifetimes::{FromFd, IntoFd};
@@ -159,6 +166,7 @@ pub trait FromHandle {
     /// # Example
     ///
     /// ```rust,no_run
+    /// # #![cfg_attr(io_lifetimes_use_std, feature(io_safety))]
     /// use std::fs::File;
     /// # use std::io;
     /// use io_lifetimes::{FromHandle, IntoHandle, OwnedHandle};
@@ -176,6 +184,7 @@ pub trait FromHandle {
     /// # Example
     ///
     /// ```rust,no_run
+    /// # #![cfg_attr(io_lifetimes_use_std, feature(io_safety))]
     /// use std::fs::File;
     /// # use std::io;
     /// use io_lifetimes::{FromHandle, IntoHandle};
@@ -207,613 +216,5 @@ pub trait FromSocket {
         Self: Sized,
     {
         Self::from_socket(into_owned.into_socket())
-    }
-}
-
-#[cfg(any(unix, target_os = "wasi"))]
-impl AsFd for BorrowedFd<'_> {
-    #[inline]
-    fn as_fd(&self) -> BorrowedFd<'_> {
-        unsafe { BorrowedFd::borrow_raw_fd(self.as_raw_fd()) }
-    }
-}
-
-#[cfg(windows)]
-impl AsHandle for BorrowedHandle<'_> {
-    #[inline]
-    fn as_handle(&self) -> BorrowedHandle<'_> {
-        unsafe { BorrowedHandle::borrow_raw_handle(self.as_raw_handle()) }
-    }
-}
-
-#[cfg(windows)]
-impl AsSocket for BorrowedSocket<'_> {
-    #[inline]
-    fn as_socket(&self) -> BorrowedSocket<'_> {
-        unsafe { BorrowedSocket::borrow_raw_socket(self.as_raw_socket()) }
-    }
-}
-
-#[cfg(any(unix, target_os = "wasi"))]
-impl AsFd for OwnedFd {
-    #[inline]
-    fn as_fd(&self) -> BorrowedFd<'_> {
-        unsafe { BorrowedFd::borrow_raw_fd(self.as_raw_fd()) }
-    }
-}
-
-#[cfg(windows)]
-impl AsHandle for OwnedHandle {
-    #[inline]
-    fn as_handle(&self) -> BorrowedHandle<'_> {
-        unsafe { BorrowedHandle::borrow_raw_handle(self.as_raw_handle()) }
-    }
-}
-
-#[cfg(windows)]
-impl AsSocket for OwnedSocket {
-    #[inline]
-    fn as_socket(&self) -> BorrowedSocket<'_> {
-        unsafe { BorrowedSocket::borrow_raw_socket(self.as_raw_socket()) }
-    }
-}
-
-#[cfg(any(unix, target_os = "wasi"))]
-impl IntoFd for OwnedFd {
-    #[inline]
-    fn into_fd(self) -> OwnedFd {
-        unsafe { Self::from_raw_fd(self.into_raw_fd()) }
-    }
-}
-
-#[cfg(windows)]
-impl IntoHandle for OwnedHandle {
-    #[inline]
-    fn into_handle(self) -> OwnedHandle {
-        unsafe { Self::from_raw_handle(self.into_raw_handle()) }
-    }
-}
-
-#[cfg(windows)]
-impl IntoSocket for OwnedSocket {
-    #[inline]
-    fn into_socket(self) -> OwnedSocket {
-        unsafe { Self::from_raw_socket(self.into_raw_socket()) }
-    }
-}
-
-#[cfg(any(unix, target_os = "wasi"))]
-impl FromFd for OwnedFd {
-    #[inline]
-    fn from_fd(owned: OwnedFd) -> Self {
-        unsafe { Self::from_raw_fd(owned.into_raw_fd()) }
-    }
-}
-
-#[cfg(windows)]
-impl FromHandle for OwnedHandle {
-    #[inline]
-    fn from_handle(owned: OwnedHandle) -> Self {
-        unsafe { Self::from_raw_handle(owned.into_raw_handle()) }
-    }
-}
-
-#[cfg(windows)]
-impl FromSocket for OwnedSocket {
-    #[inline]
-    fn from_socket(owned: OwnedSocket) -> Self {
-        unsafe { Self::from_raw_socket(owned.into_raw_socket()) }
-    }
-}
-
-#[cfg(windows)]
-impl FromHandle for HandleOrInvalid {
-    #[inline]
-    fn from_handle(owned: OwnedHandle) -> Self {
-        unsafe { Self::from_raw_handle(owned.into_raw_handle()) }
-    }
-}
-
-#[cfg(any(unix, target_os = "wasi"))]
-impl AsFd for std::fs::File {
-    #[inline]
-    fn as_fd(&self) -> BorrowedFd<'_> {
-        unsafe { BorrowedFd::borrow_raw_fd(self.as_raw_fd()) }
-    }
-}
-
-#[cfg(windows)]
-impl AsHandle for std::fs::File {
-    #[inline]
-    fn as_handle(&self) -> BorrowedHandle<'_> {
-        unsafe { BorrowedHandle::borrow_raw_handle(self.as_raw_handle()) }
-    }
-}
-
-#[cfg(any(unix, target_os = "wasi"))]
-impl IntoFd for std::fs::File {
-    #[inline]
-    fn into_fd(self) -> OwnedFd {
-        unsafe { OwnedFd::from_raw_fd(self.into_raw_fd()) }
-    }
-}
-
-#[cfg(windows)]
-impl IntoHandle for std::fs::File {
-    #[inline]
-    fn into_handle(self) -> OwnedHandle {
-        unsafe { OwnedHandle::from_raw_handle(self.into_raw_handle()) }
-    }
-}
-
-#[cfg(any(unix, target_os = "wasi"))]
-impl FromFd for std::fs::File {
-    #[inline]
-    fn from_fd(owned: OwnedFd) -> Self {
-        unsafe { Self::from_raw_fd(owned.into_raw_fd()) }
-    }
-}
-
-#[cfg(windows)]
-impl FromHandle for std::fs::File {
-    #[inline]
-    fn from_handle(owned: OwnedHandle) -> Self {
-        unsafe { Self::from_raw_handle(owned.into_raw_handle()) }
-    }
-}
-
-#[cfg(any(unix, target_os = "wasi"))]
-impl AsFd for std::net::TcpStream {
-    #[inline]
-    fn as_fd(&self) -> BorrowedFd<'_> {
-        unsafe { BorrowedFd::borrow_raw_fd(self.as_raw_fd()) }
-    }
-}
-
-#[cfg(windows)]
-impl AsSocket for std::net::TcpStream {
-    #[inline]
-    fn as_socket(&self) -> BorrowedSocket<'_> {
-        unsafe { BorrowedSocket::borrow_raw_socket(self.as_raw_socket()) }
-    }
-}
-
-#[cfg(any(unix, target_os = "wasi"))]
-impl IntoFd for std::net::TcpStream {
-    #[inline]
-    fn into_fd(self) -> OwnedFd {
-        unsafe { OwnedFd::from_raw_fd(self.into_raw_fd()) }
-    }
-}
-
-#[cfg(windows)]
-impl IntoSocket for std::net::TcpStream {
-    #[inline]
-    fn into_socket(self) -> OwnedSocket {
-        unsafe { OwnedSocket::from_raw_socket(self.into_raw_socket()) }
-    }
-}
-
-#[cfg(any(unix, target_os = "wasi"))]
-impl FromFd for std::net::TcpStream {
-    #[inline]
-    fn from_fd(owned: OwnedFd) -> Self {
-        unsafe { Self::from_raw_fd(owned.into_raw_fd()) }
-    }
-}
-
-#[cfg(windows)]
-impl FromSocket for std::net::TcpStream {
-    #[inline]
-    fn from_socket(owned: OwnedSocket) -> Self {
-        unsafe { Self::from_raw_socket(owned.into_raw_socket()) }
-    }
-}
-
-#[cfg(any(unix, target_os = "wasi"))]
-impl AsFd for std::net::TcpListener {
-    #[inline]
-    fn as_fd(&self) -> BorrowedFd<'_> {
-        unsafe { BorrowedFd::borrow_raw_fd(self.as_raw_fd()) }
-    }
-}
-
-#[cfg(windows)]
-impl AsSocket for std::net::TcpListener {
-    #[inline]
-    fn as_socket(&self) -> BorrowedSocket<'_> {
-        unsafe { BorrowedSocket::borrow_raw_socket(self.as_raw_socket()) }
-    }
-}
-
-#[cfg(any(unix, target_os = "wasi"))]
-impl IntoFd for std::net::TcpListener {
-    #[inline]
-    fn into_fd(self) -> OwnedFd {
-        unsafe { OwnedFd::from_raw_fd(self.into_raw_fd()) }
-    }
-}
-
-#[cfg(windows)]
-impl IntoSocket for std::net::TcpListener {
-    #[inline]
-    fn into_socket(self) -> OwnedSocket {
-        unsafe { OwnedSocket::from_raw_socket(self.into_raw_socket()) }
-    }
-}
-
-#[cfg(any(unix, target_os = "wasi"))]
-impl FromFd for std::net::TcpListener {
-    #[inline]
-    fn from_fd(owned: OwnedFd) -> Self {
-        unsafe { Self::from_raw_fd(owned.into_raw_fd()) }
-    }
-}
-
-#[cfg(windows)]
-impl FromSocket for std::net::TcpListener {
-    #[inline]
-    fn from_socket(owned: OwnedSocket) -> Self {
-        unsafe { Self::from_raw_socket(owned.into_raw_socket()) }
-    }
-}
-
-#[cfg(any(unix, target_os = "wasi"))]
-impl AsFd for std::net::UdpSocket {
-    #[inline]
-    fn as_fd(&self) -> BorrowedFd<'_> {
-        unsafe { BorrowedFd::borrow_raw_fd(self.as_raw_fd()) }
-    }
-}
-
-#[cfg(windows)]
-impl AsSocket for std::net::UdpSocket {
-    #[inline]
-    fn as_socket(&self) -> BorrowedSocket {
-        unsafe { BorrowedSocket::borrow_raw_socket(self.as_raw_socket()) }
-    }
-}
-
-#[cfg(any(unix, target_os = "wasi"))]
-impl IntoFd for std::net::UdpSocket {
-    #[inline]
-    fn into_fd(self) -> OwnedFd {
-        unsafe { OwnedFd::from_raw_fd(self.into_raw_fd()) }
-    }
-}
-
-#[cfg(windows)]
-impl IntoSocket for std::net::UdpSocket {
-    #[inline]
-    fn into_socket(self) -> OwnedSocket {
-        unsafe { OwnedSocket::from_raw_socket(self.into_raw_socket()) }
-    }
-}
-
-#[cfg(any(unix, target_os = "wasi"))]
-impl FromFd for std::net::UdpSocket {
-    #[inline]
-    fn from_fd(owned: OwnedFd) -> Self {
-        unsafe { Self::from_raw_fd(owned.into_raw_fd()) }
-    }
-}
-
-#[cfg(windows)]
-impl FromSocket for std::net::UdpSocket {
-    #[inline]
-    fn from_socket(owned: OwnedSocket) -> Self {
-        unsafe { Self::from_raw_socket(owned.into_raw_socket()) }
-    }
-}
-
-#[cfg(any(unix, target_os = "wasi"))]
-impl AsFd for std::io::Stdin {
-    #[inline]
-    fn as_fd(&self) -> BorrowedFd<'_> {
-        unsafe { BorrowedFd::borrow_raw_fd(self.as_raw_fd()) }
-    }
-}
-
-#[cfg(windows)]
-impl AsHandle for std::io::Stdin {
-    #[inline]
-    fn as_handle(&self) -> BorrowedHandle<'_> {
-        unsafe { BorrowedHandle::borrow_raw_handle(self.as_raw_handle()) }
-    }
-}
-
-#[cfg(any(unix, target_os = "wasi"))]
-impl<'a> AsFd for std::io::StdinLock<'a> {
-    #[inline]
-    fn as_fd(&self) -> BorrowedFd<'_> {
-        unsafe { BorrowedFd::borrow_raw_fd(self.as_raw_fd()) }
-    }
-}
-
-#[cfg(windows)]
-impl<'a> AsHandle for std::io::StdinLock<'a> {
-    #[inline]
-    fn as_handle(&self) -> BorrowedHandle<'_> {
-        unsafe { BorrowedHandle::borrow_raw_handle(self.as_raw_handle()) }
-    }
-}
-
-#[cfg(any(unix, target_os = "wasi"))]
-impl AsFd for std::io::Stdout {
-    #[inline]
-    fn as_fd(&self) -> BorrowedFd<'_> {
-        unsafe { BorrowedFd::borrow_raw_fd(self.as_raw_fd()) }
-    }
-}
-
-#[cfg(windows)]
-impl AsHandle for std::io::Stdout {
-    #[inline]
-    fn as_handle(&self) -> BorrowedHandle<'_> {
-        unsafe { BorrowedHandle::borrow_raw_handle(self.as_raw_handle()) }
-    }
-}
-
-#[cfg(any(unix, target_os = "wasi"))]
-impl<'a> AsFd for std::io::StdoutLock<'a> {
-    #[inline]
-    fn as_fd(&self) -> BorrowedFd<'_> {
-        unsafe { BorrowedFd::borrow_raw_fd(self.as_raw_fd()) }
-    }
-}
-
-#[cfg(windows)]
-impl<'a> AsHandle for std::io::StdoutLock<'a> {
-    #[inline]
-    fn as_handle(&self) -> BorrowedHandle<'_> {
-        unsafe { BorrowedHandle::borrow_raw_handle(self.as_raw_handle()) }
-    }
-}
-
-#[cfg(any(unix, target_os = "wasi"))]
-impl AsFd for std::io::Stderr {
-    #[inline]
-    fn as_fd(&self) -> BorrowedFd<'_> {
-        unsafe { BorrowedFd::borrow_raw_fd(self.as_raw_fd()) }
-    }
-}
-
-#[cfg(windows)]
-impl AsHandle for std::io::Stderr {
-    #[inline]
-    fn as_handle(&self) -> BorrowedHandle<'_> {
-        unsafe { BorrowedHandle::borrow_raw_handle(self.as_raw_handle()) }
-    }
-}
-
-#[cfg(any(unix, target_os = "wasi"))]
-impl<'a> AsFd for std::io::StderrLock<'a> {
-    #[inline]
-    fn as_fd(&self) -> BorrowedFd<'_> {
-        unsafe { BorrowedFd::borrow_raw_fd(self.as_raw_fd()) }
-    }
-}
-
-#[cfg(windows)]
-impl<'a> AsHandle for std::io::StderrLock<'a> {
-    #[inline]
-    fn as_handle(&self) -> BorrowedHandle<'_> {
-        unsafe { BorrowedHandle::borrow_raw_handle(self.as_raw_handle()) }
-    }
-}
-
-#[cfg(unix)]
-impl AsFd for std::process::ChildStdin {
-    #[inline]
-    fn as_fd(&self) -> BorrowedFd<'_> {
-        unsafe { BorrowedFd::borrow_raw_fd(self.as_raw_fd()) }
-    }
-}
-
-#[cfg(windows)]
-impl AsHandle for std::process::ChildStdin {
-    #[inline]
-    fn as_handle(&self) -> BorrowedHandle<'_> {
-        unsafe { BorrowedHandle::borrow_raw_handle(self.as_raw_handle()) }
-    }
-}
-
-#[cfg(unix)]
-impl IntoFd for std::process::ChildStdin {
-    #[inline]
-    fn into_fd(self) -> OwnedFd {
-        unsafe { OwnedFd::from_raw_fd(self.into_raw_fd()) }
-    }
-}
-
-#[cfg(windows)]
-impl IntoHandle for std::process::ChildStdin {
-    #[inline]
-    fn into_handle(self) -> OwnedHandle {
-        unsafe { OwnedHandle::from_raw_handle(self.into_raw_handle()) }
-    }
-}
-
-#[cfg(unix)]
-impl AsFd for std::process::ChildStdout {
-    #[inline]
-    fn as_fd(&self) -> BorrowedFd<'_> {
-        unsafe { BorrowedFd::borrow_raw_fd(self.as_raw_fd()) }
-    }
-}
-
-#[cfg(windows)]
-impl AsHandle for std::process::ChildStdout {
-    #[inline]
-    fn as_handle(&self) -> BorrowedHandle<'_> {
-        unsafe { BorrowedHandle::borrow_raw_handle(self.as_raw_handle()) }
-    }
-}
-
-#[cfg(unix)]
-impl IntoFd for std::process::ChildStdout {
-    #[inline]
-    fn into_fd(self) -> OwnedFd {
-        unsafe { OwnedFd::from_raw_fd(self.into_raw_fd()) }
-    }
-}
-
-#[cfg(windows)]
-impl IntoHandle for std::process::ChildStdout {
-    #[inline]
-    fn into_handle(self) -> OwnedHandle {
-        unsafe { OwnedHandle::from_raw_handle(self.into_raw_handle()) }
-    }
-}
-
-#[cfg(unix)]
-impl AsFd for std::process::ChildStderr {
-    #[inline]
-    fn as_fd(&self) -> BorrowedFd<'_> {
-        unsafe { BorrowedFd::borrow_raw_fd(self.as_raw_fd()) }
-    }
-}
-
-#[cfg(windows)]
-impl AsHandle for std::process::ChildStderr {
-    #[inline]
-    fn as_handle(&self) -> BorrowedHandle<'_> {
-        unsafe { BorrowedHandle::borrow_raw_handle(self.as_raw_handle()) }
-    }
-}
-
-#[cfg(unix)]
-impl IntoFd for std::process::ChildStderr {
-    #[inline]
-    fn into_fd(self) -> OwnedFd {
-        unsafe { OwnedFd::from_raw_fd(self.into_raw_fd()) }
-    }
-}
-
-#[cfg(windows)]
-impl IntoHandle for std::process::ChildStderr {
-    #[inline]
-    fn into_handle(self) -> OwnedHandle {
-        unsafe { OwnedHandle::from_raw_handle(self.into_raw_handle()) }
-    }
-}
-
-#[cfg(unix)]
-impl FromFd for std::process::Stdio {
-    #[inline]
-    fn from_fd(owned: OwnedFd) -> Self {
-        unsafe { Self::from_raw_fd(owned.into_raw_fd()) }
-    }
-}
-
-#[cfg(windows)]
-impl FromHandle for std::process::Stdio {
-    #[inline]
-    fn from_handle(owned: OwnedHandle) -> Self {
-        unsafe { Self::from_raw_handle(owned.into_raw_handle()) }
-    }
-}
-
-#[cfg(windows)]
-impl AsHandle for std::process::Child {
-    #[inline]
-    fn as_handle(&self) -> BorrowedHandle<'_> {
-        unsafe { BorrowedHandle::borrow_raw_handle(self.as_raw_handle()) }
-    }
-}
-
-#[cfg(windows)]
-impl IntoHandle for std::process::Child {
-    #[inline]
-    fn into_handle(self) -> OwnedHandle {
-        unsafe { OwnedHandle::from_raw_handle(self.into_raw_handle()) }
-    }
-}
-
-#[cfg(unix)]
-impl AsFd for std::os::unix::net::UnixStream {
-    #[inline]
-    fn as_fd(&self) -> BorrowedFd<'_> {
-        unsafe { BorrowedFd::borrow_raw_fd(self.as_raw_fd()) }
-    }
-}
-
-#[cfg(unix)]
-impl IntoFd for std::os::unix::net::UnixStream {
-    #[inline]
-    fn into_fd(self) -> OwnedFd {
-        unsafe { OwnedFd::from_raw_fd(self.into_raw_fd()) }
-    }
-}
-
-#[cfg(unix)]
-impl FromFd for std::os::unix::net::UnixStream {
-    #[inline]
-    fn from_fd(owned: OwnedFd) -> Self {
-        unsafe { Self::from_raw_fd(owned.into_raw_fd()) }
-    }
-}
-
-#[cfg(unix)]
-impl AsFd for std::os::unix::net::UnixListener {
-    #[inline]
-    fn as_fd(&self) -> BorrowedFd<'_> {
-        unsafe { BorrowedFd::borrow_raw_fd(self.as_raw_fd()) }
-    }
-}
-
-#[cfg(unix)]
-impl IntoFd for std::os::unix::net::UnixListener {
-    #[inline]
-    fn into_fd(self) -> OwnedFd {
-        unsafe { OwnedFd::from_raw_fd(self.into_raw_fd()) }
-    }
-}
-
-#[cfg(unix)]
-impl FromFd for std::os::unix::net::UnixListener {
-    #[inline]
-    fn from_fd(owned: OwnedFd) -> Self {
-        unsafe { Self::from_raw_fd(owned.into_raw_fd()) }
-    }
-}
-
-#[cfg(unix)]
-impl AsFd for std::os::unix::net::UnixDatagram {
-    #[inline]
-    fn as_fd(&self) -> BorrowedFd<'_> {
-        unsafe { BorrowedFd::borrow_raw_fd(self.as_raw_fd()) }
-    }
-}
-
-#[cfg(unix)]
-impl IntoFd for std::os::unix::net::UnixDatagram {
-    #[inline]
-    fn into_fd(self) -> OwnedFd {
-        unsafe { OwnedFd::from_raw_fd(self.into_raw_fd()) }
-    }
-}
-
-#[cfg(unix)]
-impl FromFd for std::os::unix::net::UnixDatagram {
-    #[inline]
-    fn from_fd(owned: OwnedFd) -> Self {
-        unsafe { Self::from_raw_fd(owned.into_raw_fd()) }
-    }
-}
-
-#[cfg(windows)]
-impl<T> AsHandle for std::thread::JoinHandle<T> {
-    #[inline]
-    fn as_handle(&self) -> BorrowedHandle<'_> {
-        unsafe { BorrowedHandle::borrow_raw_handle(self.as_raw_handle()) }
-    }
-}
-
-#[cfg(windows)]
-impl<T> IntoHandle for std::thread::JoinHandle<T> {
-    #[inline]
-    fn into_handle(self) -> OwnedHandle {
-        unsafe { OwnedHandle::from_raw_handle(self.into_raw_handle()) }
     }
 }
