+++ tests/niche-optimizations.rs
@@ -1,5 +1,6 @@
 #![cfg_attr(not(rustc_attrs), allow(unused_imports))]
 #![cfg_attr(target_os = "wasi", feature(wasi_ext))]
+#![cfg_attr(io_lifetimes_use_std, feature(io_safety))]
 
 use std::mem::size_of;
